python/002-connect-connection-string/requirements.txt
azure-cosmos==4.9.0
python-dotenv==1.0.1
